requirements.txt
google-genai
psycopg2-binary==2.9.10
pgvector==0.3.2
python-dotenv==1.0.1
pypdf==5.1.0
python-docx==1.1.2
